Attempt to silence unused var warnings with cuda-11.4.2-sems-gnu-10.1.0 PR build
The PR build using cuda-11.4.2-sems-gnu-10.1.0 is generating a bunch of unused
varaible warnings that are not being seen on other compilers and I can't seem
to find a way to silence all of them.

For now, I just focused on trying to make all of the warnings go away for the
translation unit:

  teuchos/core/test/MemoryManagement/Array_UnitTests.cpp

For unusesd function parameter warnings, I put in the typical:

  (void)<varname>;

statements, but that clutters up the unit tests.

A better solution is to use the C++17 [[maybe_unussed]] attribute for all of
these.  But it seems that this CUDA+GNU complier does not honor that attribute
and it is producing the unused warnings anyway. I guess there is just no hope
for a warnings-free build with this CUDA compiler :-(.

The other irritating thing is that this compiler is generating this warning
for RAII objects that are, by design, supposed to never get "used".
(I.e. their only usage is to do something on construction and on implicit
destruction at the end of the scope and no code ever mentions those vars after
construction.)

I left in a couple of the [[maybe_unused]] attributes in hopes this may help
other complers to avoid these warnings (and show how to use this).

But at least clang-11.0.1 does not show any of these warnings.